* 'master' of https://github.com/giampaolo/psutil:
  pre-release
  give CREDIT to @JeremyGrosser and @getsentry for sponsorship (thanks\!)
  fix error in test_contracts.py on win
  chore: update GHA workflows (#2315)
  fix #2195 / linux: no longer print exception at import time
  Add ruff Python linter, remove flake8 (#2312)
  Add toml-sort linting tool for pyproject.toml (#2311)
  new RTD config
  new RTD config
  new RTD config
  Fix #2308, OpenBSD: Process.threads() always fail with AccessDenied.
  Win: fix OpenProcess not recognizing when proc is gone.
  refact assertProcessGone
  refact assertProcessGone
  Add RsT linter (#2292)
  Make _raise_if_pid_reused() raise NSP only if pid has been reused.
  fix DEVGUIDE.rst
  PID reusage is not checked for different set methods (#2291)

+++ psutil/__init__.py
@@ -227,8 +227,8 @@ _SENTINEL = object()
 # See: https://github.com/giampaolo/psutil/issues/564
 if (int(__version__.replace('.', '')) !=
         getattr(_psplatform.cext, 'version', None)):
-    msg = "version conflict: %r C extension module was built for another " \
-          "version of psutil" % _psplatform.cext.__file__
+    msg = "version conflict: %r C extension " % _psplatform.cext.__file__
+    msg += "module was built for another version of psutil"
     if hasattr(_psplatform.cext, 'version'):
         msg += " (%s instead of %s)" % (
             '.'.join([x for x in str(_psplatform.cext.version)]), __version__)
@@ -263,30 +263,11 @@ else:  # pragma: no cover
         return ret
 
 
-def _assert_pid_not_reused(fun):
-    """Decorator which raises NoSuchProcess in case a process is no
-    longer running or its PID has been reused.
-    """
-    @functools.wraps(fun)
-    def wrapper(self, *args, **kwargs):
-        if not self.is_running():
-            if self._pid_reused:
-                msg = "process no longer exists and its PID has been reused"
-            else:
-                msg = None
-            raise NoSuchProcess(self.pid, self._name, msg=msg)
-        return fun(self, *args, **kwargs)
-    return wrapper
-
-
 def _pprint_secs(secs):
     """Format seconds in a human readable form."""
     now = time.time()
     secs_ago = int(now - secs)
-    if secs_ago < 60 * 60 * 24:
-        fmt = "%H:%M:%S"
-    else:
-        fmt = "%Y-%m-%d %H:%M:%S"
+    fmt = "%H:%M:%S" if secs_ago < 60 * 60 * 24 else "%Y-%m-%d %H:%M:%S"
     return datetime.datetime.fromtimestamp(secs).strftime(fmt)
 
 
@@ -295,7 +276,7 @@ def _pprint_secs(secs):
 # =====================================================================
 
 
-class Process(object):
+class Process(object):  # noqa: UP004
     """Represents an OS process with the given PID.
     If PID is omitted current process PID (os.getpid()) is used.
     Raise NoSuchProcess if PID does not exist.
@@ -441,6 +422,18 @@ class Process(object):
         if self._hash is None:
             self._hash = hash(self._ident)
         return self._hash
+
+    def _raise_if_pid_reused(self):
+        """Raises NoSuchProcess in case process PID has been reused."""
+        if not self.is_running() and self._pid_reused:
+            # We may directly raise NSP in here already if PID is just
+            # not running, but I prefer NSP to be raised naturally by
+            # the actual Process API call. This way unit tests will tell
+            # us if the API is broken (aka don't raise NSP when it
+            # should). We also remain consistent with all other "get"
+            # APIs which don't use _raise_if_pid_reused().
+            msg = "process no longer exists and its PID has been reused"
+            raise NoSuchProcess(self.pid, self._name, msg=msg)
 
     @property
     def pid(self):
@@ -627,6 +620,7 @@ class Process(object):
 
         # XXX should we check creation time here rather than in
         # Process.parent()?
+        self._raise_if_pid_reused()
         if POSIX:
             return self._proc.ppid()
         else:  # pragma: no cover
@@ -750,8 +744,7 @@ class Process(object):
         if value is None:
             return self._proc.nice_get()
         else:
-            if not self.is_running():
-                raise NoSuchProcess(self.pid, self._name)
+            self._raise_if_pid_reused()
             self._proc.nice_set(value)
 
     if POSIX:
@@ -813,6 +806,7 @@ class Process(object):
                     raise ValueError("'ioclass' argument must be specified")
                 return self._proc.ionice_get()
             else:
+                self._raise_if_pid_reused()
                 return self._proc.ionice_set(ioclass, value)
 
     # Linux / FreeBSD only
@@ -828,6 +822,8 @@ class Process(object):
             See "man prlimit" for further info.
             Available on Linux and FreeBSD only.
             """
+            if limits is not None:
+                self._raise_if_pid_reused()
             return self._proc.rlimit(resource, limits)
 
     # Windows, Linux and FreeBSD only
@@ -844,6 +840,7 @@ class Process(object):
             if cpus is None:
                 return sorted(set(self._proc.cpu_affinity_get()))
             else:
+                self._raise_if_pid_reused()
                 if not cpus:
                     if hasattr(self._proc, "_get_eligible_cpus"):
                         cpus = self._proc._get_eligible_cpus()
@@ -869,7 +866,8 @@ class Process(object):
 
         def environ(self):
             """The environment variables of the process as a dict.  Note: this
-            might not reflect changes made after the process started.  """
+            might not reflect changes made after the process started.
+            """
             return self._proc.environ()
 
     if WINDOWS:
@@ -900,7 +898,6 @@ class Process(object):
             """
             return self._proc.threads()
 
-    @_assert_pid_not_reused
     def children(self, recursive=False):
         """Return the children of this process as a list of Process
         instances, pre-emptively checking whether PID has been reused.
@@ -927,6 +924,7 @@ class Process(object):
         process Y won't be listed as the reference to process A
         is lost.
         """
+        self._raise_if_pid_reused()
         ppid_map = _ppid_map()
         ret = []
         if not recursive:
@@ -1197,6 +1195,7 @@ class Process(object):
     if POSIX:
         def _send_signal(self, sig):
             assert not self.pid < 0, self.pid
+            self._raise_if_pid_reused()
             if self.pid == 0:
                 # see "man 2 kill"
                 raise ValueError(
@@ -1216,7 +1215,6 @@ class Process(object):
             except PermissionError:
                 raise AccessDenied(self.pid, self._name)
 
-    @_assert_pid_not_reused
     def send_signal(self, sig):
         """Send a signal *sig* to process pre-emptively checking
         whether PID has been reused (see signal module constants) .
@@ -1226,9 +1224,12 @@ class Process(object):
         if POSIX:
             self._send_signal(sig)
         else:  # pragma: no cover
+            self._raise_if_pid_reused()
+            if sig != signal.SIGTERM and not self.is_running():
+                msg = "process no longer exists"
+                raise NoSuchProcess(self.pid, self._name, msg=msg)
             self._proc.send_signal(sig)
 
-    @_assert_pid_not_reused
     def suspend(self):
         """Suspend process execution with SIGSTOP pre-emptively checking
         whether PID has been reused.
@@ -1237,9 +1238,9 @@ class Process(object):
         if POSIX:
             self._send_signal(signal.SIGSTOP)
         else:  # pragma: no cover
+            self._raise_if_pid_reused()
             self._proc.suspend()
 
-    @_assert_pid_not_reused
     def resume(self):
         """Resume process execution with SIGCONT pre-emptively checking
         whether PID has been reused.
@@ -1248,9 +1249,9 @@ class Process(object):
         if POSIX:
             self._send_signal(signal.SIGCONT)
         else:  # pragma: no cover
+            self._raise_if_pid_reused()
             self._proc.resume()
 
-    @_assert_pid_not_reused
     def terminate(self):
         """Terminate the process with SIGTERM pre-emptively checking
         whether PID has been reused.
@@ -1259,9 +1260,9 @@ class Process(object):
         if POSIX:
             self._send_signal(signal.SIGTERM)
         else:  # pragma: no cover
+            self._raise_if_pid_reused()
             self._proc.kill()
 
-    @_assert_pid_not_reused
     def kill(self):
         """Kill the current process with SIGKILL pre-emptively checking
         whether PID has been reused.
@@ -1269,6 +1270,7 @@ class Process(object):
         if POSIX:
             self._send_signal(signal.SIGKILL)
         else:  # pragma: no cover
+            self._raise_if_pid_reused()
             self._proc.kill()
 
     def wait(self, timeout=None):
@@ -1329,7 +1331,7 @@ class Popen(Process):
       >>> p.username()
       'giampaolo'
       >>> p.communicate()
-      ('hi\n', None)
+      ('hi', None)
       >>> p.terminate()
       >>> p.wait(timeout=2)
       0
@@ -1380,7 +1382,7 @@ class Popen(Process):
     def wait(self, timeout=None):
         if self.__subproc.returncode is not None:
             return self.__subproc.returncode
-        ret = super(Popen, self).wait(timeout)
+        ret = super(Popen, self).wait(timeout)  # noqa
         self.__subproc.returncode = ret
         return ret
 
@@ -2196,7 +2198,7 @@ def net_if_addrs():
                 if WINDOWS and fam == -1:
                     fam = _psplatform.AF_LINK
                 elif (hasattr(_psplatform, "AF_LINK") and
-                        _psplatform.AF_LINK == fam):
+                        fam == _psplatform.AF_LINK):
                     # Linux defines AF_LINK as an alias for AF_PACKET.
                     # We re-set the family here so that repr(family)
                     # will show AF_LINK rather than AF_PACKET
@@ -2421,7 +2423,7 @@ def test():  # pragma: no cover
 
 del memoize_when_activated, division
 if sys.version_info[0] < 3:
-    del num, x
+    del num, x  # noqa
 
 if __name__ == "__main__":
     test()
